requirements.txt
pystac-client>=0.7.0
planetary-computer>=0.5.0
shapely>=2.0.0
rasterio>=1.3.0
requests>=2.31.0
boto3>=1.34.0
openai>=1.0.0
pandas>=2.0.0
pillow>=10.0.0
openpyxl>=3.0.0  # Required for Excel support in pandas